e2e/requirements.txt
pytest
requests
w3storage
python-dotenv